refactor: move deploymentConfig into internal folder
